Bugfix in full time extraction: numberOfGrids back to float

+++ tracking/modules/trackTimeExtraction/include/FullGridTrackTimeExtractionModule.h
@@ -63,7 +63,7 @@ namespace Belle2 {
     /// Module Parameter: Maximal shift which is allowed.
     double m_param_maximalT0Shift = 30;
     /// Module Parameter: Number of shifts to try out.
-    unsigned int m_param_numberOfGrids = 5;
+    float m_param_numberOfGrids = 5.0;
     /// Module Parameter: Whether to replace an existing time estimation or not.
     bool m_param_overwriteExistingEstimation = true;
     /// Module Parameter: Use this as sigma t0.
